# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/employees": {
      "post": {
        "operationId": "EmployeeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeDto"
              },
              "examples": {
                "default": {
                  "summary": "New employee",
                  "value": {
                    "name": "Ahmed Ali",
                    "email": "ahmed@acme.com",
                    "basicSalary": 5000,
                    "employmentType": "PERMANENT",
                    "salaryBasis": "MONTHLY",
                    "isGosiRegistered": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Employees"
        ]
      },
      "get": {
        "operationId": "EmployeeController_findAll",
        "parameters": [
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Employees"
        ]
      }
    },
    "/employees/{id}": {
      "get": {
        "operationId": "EmployeeController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Employees"
        ]
      },
      "patch": {
        "operationId": "EmployeeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeDto"
              },
              "examples": {
                "default": {
                  "summary": "Update",
                  "value": {
                    "basicSalary": 6000,
                    "isActive": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Employees"
        ]
      }
    },
    "/employees/{employeeId}/documents": {
      "post": {
        "operationId": "DocumentController_create",
        "parameters": [
          {
            "name": "employeeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentDto"
              },
              "examples": {
                "default": {
                  "summary": "Add document",
                  "value": {
                    "type": "NATIONAL_ID",
                    "expiryDate": "2027-12-31",
                    "documentNumber": "1234567890",
                    "fileUrl": "https://files.example.com/id.pdf"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Documents"
        ]
      },
      "get": {
        "operationId": "DocumentController_findAll",
        "parameters": [
          {
            "name": "employeeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Documents"
        ]
      }
    },
    "/documents/expiring": {
      "get": {
        "operationId": "DocumentController_findExpiring",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Documents"
        ]
      }
    },
    "/documents/{id}": {
      "delete": {
        "operationId": "DocumentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Documents"
        ]
      }
    },
...
    "/attendance": {
      "post": {
        "operationId": "AttendanceController_upsert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertAttendanceDto"
              },
              "examples": {
                "default": {
                  "summary": "Manual entry (needs shift on employee)",
                  "value": {
                    "employeeId": "employee-uuid",
                    "date": "2026-06-29",
                    "checkIn": "2026-06-29T08:05:00+03:00",
                    "checkOut": "2026-06-29T17:30:00+03:00"
                  }
                },
                "leave": {
                  "summary": "Mark a leave day",
                  "value": {
                    "employeeId": "employee-uuid",
                    "date": "2026-06-29",
                    "status": "LEAVE"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Attendance"
        ]
      },
      "get": {
        "operationId": "AttendanceController_findAll",
        "parameters": [
          {
            "name": "employeeId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PRESENT",
                "ABSENT",
                "LEAVE"
              ]
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "Inclusive lower bound (date).",
            "schema": {
              "example": "2026-06-01",
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "Inclusive upper bound (date).",
            "schema": {
              "example": "2026-06-30",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Attendance"
        ]
      }
    },
...
    "/loans": {
      "get": {
        "operationId": "LoanController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "PAID_OFF"
              ]
            }
          },
          {
            "name": "employeeId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Loans"
        ]
      }
    },
...
    "/ess/me": {
      "get": {
        "operationId": "EssController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ESS"
        ]
      }
    },
    "/ess/salary-components": {
      "get": {
        "operationId": "EssController_salaryComponents",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ESS"
        ]
      }
    },
    "/ess/documents": {
      "get": {
        "operationId": "EssController_documents",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ESS"
        ]
      }
    },
    "/ess/attendance": {
      "get": {
        "operationId": "EssController_attendance",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ESS"
        ]
      }
    },
    "/ess/loans": {
      "get": {
        "operationId": "EssController_loans",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ESS"
        ]
      }
    },
    "/ess/payslips": {
      "get": {
        "operationId": "EssController_payslips",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ESS"
        ]
      }
    },
    "/ess/payslips/{id}": {
      "get": {
        "operationId": "EssController_payslip",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ESS"
        ]
      }
    },
...
  },
  "info": {
    "title": "HR System API",
    "description": "Multi-tenant Payroll & HR Management API (Phases 1–6). Import into APIDog via live URL http://localhost:3004/docs-json or backend/postman_collection.json. Swagger UI: http://localhost:3004/docs",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3004/api",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    },
    "schemas": {
...
      "CreateEmployeeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ahmed Ali",
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "example": "ahmed@acme.com",
            "format": "email",
            "description": "Login email for the auto-created portal account (unique)."
          },
          "basicSalary": {
            "type": "number",
            "example": 5000,
            "minimum": 0,
            "description": "Basic salary"
          },
          "employmentType": {
            "type": "string",
            "enum": [
              "PERMANENT",
              "TEMPORARY",
              "CONTRACT",
              "PROBATION"
            ],
            "example": "PERMANENT"
          },
          "salaryBasis": {
            "type": "string",
            "enum": [
              "MONTHLY",
              "DAILY",
              "HOURLY"
            ],
            "example": "MONTHLY"
          },
          "shiftId": {
            "type": "string",
            "format": "uuid",
            "description": "Optional shift to assign (must belong to your company)."
          },
          "isGosiRegistered": {
            "type": "boolean",
            "example": false
          },
          "gosiNumber": {
            "type": "string",
            "example": "123456789",
            "maxLength": 50
          }
        },
        "required": [
          "name",
          "email",
          "basicSalary"
        ]
      },
      "UpdateEmployeeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ahmed Ali",
            "maxLength": 120
          },
          "basicSalary": {
            "type": "number",
            "example": 6000,
            "minimum": 0
          },
          "employmentType": {
            "type": "string",
            "enum": [
              "PERMANENT",
              "TEMPORARY",
              "CONTRACT",
              "PROBATION"
            ]
          },
          "salaryBasis": {
            "type": "string",
            "enum": [
              "MONTHLY",
              "DAILY",
              "HOURLY"
            ]
          },
          "shiftId": {
            "type": "string",
            "format": "uuid"
          },
          "isActive": {
            "type": "boolean",
            "example": false,
            "description": "Set to false to deactivate the employee (resignation)."
          },
          "isGosiRegistered": {
            "type": "boolean",
            "example": true
          },
          "gosiNumber": {
            "type": "string",
            "example": "123456789",
            "maxLength": 50
          }
        }
      },
      "CreateDocumentDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "NATIONAL_ID",
              "RESIDENCY",
              "PASSPORT",
              "CONTRACT",
              "CERTIFICATE",
              "OTHER"
            ],
            "example": "NATIONAL_ID"
          },
          "expiryDate": {
            "type": "string",
            "example": "2027-12-31",
            "description": "Expiry date (drives renewal reminders)."
          },
          "fileUrl": {
            "type": "string",
            "example": "https://files.example.com/id.pdf"
          },
          "documentNumber": {
            "type": "string",
            "example": "1234567890",
            "maxLength": 50
          }
        },
        "required": [
          "type"
        ]
      },
...
      "UpsertAttendanceDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "example": "2026-06-29",
            "description": "Calendar day (tenant-local)."
          },
          "shiftId": {
            "type": "string",
            "format": "uuid"
          },
          "checkIn": {
            "type": "string",
            "example": "2026-06-29T08:05:00+03:00"
          },
          "checkOut": {
            "type": "string",
            "example": "2026-06-29T17:30:00+03:00"
          },
          "status": {
            "type": "string",
            "enum": [
              "PRESENT",
              "ABSENT",
              "LEAVE"
            ],
            "example": "PRESENT"
          },
          "delayMinutes": {
            "type": "number",
            "minimum": 0,
            "description": "Manual override of derived delay."
          },
          "overtimeHours": {
            "type": "number",
            "minimum": 0,
            "description": "Manual override of derived overtime."
          }
        },
        "required": [
          "employeeId",
          "date"
        ]
      },
...
    }
  }
}